+++ p4spec-rust/tests/sim-plugin/ebpf/mod.rs
@@ -129,7 +129,7 @@ fn test_native_ebpf_micro_fixture_packets() {
                     txs.extend(state.txs.iter().map(|tx| (tx.port, tx.packet.clone())));
                 }
                 Statement::Expect { port, packet_expected: Some(packet), exact: _ } => {
-                    txs_expect.push((port.parse::<i64>().unwrap(), packet))
+                    txs_expect.push((port.parse::<usize>().unwrap(), packet))
                 }
                 _ => panic!("micro fixture contains packet and expectation statements"),
             }
